SERVICES: SERVICES contain 1 class and 1 interface (CustomerService and ICustomerService.) Interface has been implemented.
ICustomerService declares what methods CustomerService must implement, without saying how. It states that we need to include information about customer, and also needs to include error handling in case of any error occuring.

CustomerService currently includes: fullfilling the ICustomerService contract.
Also includes creating customer with GUID and adding all of the information that the customer fills in, and sends this to CustomerRepo (where the Json file is located.)

REPOS:
REPOS contain 1 class and 1 interface (CustomerRepo and ICustomerRepo.)

ICustomerRepo defines what CustomerService must do. In this case the CustomerRepo needs to 1. retrieve all customers and return them as a list and 2. actually saving the list containing all the customer information, and confirm if it was succesful or not with a bool.

CustomerRepo currently includes: fulfilling the ICustomerRepo contract by handling of all the data, including saving a list with all the customer information down to a json file. CustomerRepo stores the data, while CustomerService sends data to it to save, and requests data from it to read.

MODELS:
MODELS contain 2 classes (AddressModel and CustomerModel.)

AddressModel includes: information on the customers Address, postal code and city.

CustomerModel contains fields/properties representing a customer: Customer Id, Customer name, customer email, customer phone number.

(CustomerModel and CustomerAddress has been separated in 2 sections in order to follow Single Responsibility Principle.)

CONTROLLERS:
The MenuController handles the interaction with the user. It contains a menu with options to 1. view current customers 2. to add new customers and 3. the option to exit the whole menu. More options will be added as the project grows.